Link logs to traces for adservice (java)
This config change dumps the trace and span IDs into logs for adservice.
The Java OTel SDK inject trace context into MDC, and we can grab that
via log4j. I think there are ways to do this for other services, as
well! And in general, we need to have sane logging for all of our
services, because we're mostly just spewing debug logs into tempo
right now.