Link traces back to logs (all services)
This allows us to link from a trace back to "relevant logs" for any span
and any service. To do so we:
- sanitize and munge the docker container name into a `service_name`
  label for loki
- insert a `service_name` attribute on every span, copying it from the
  `service.name` resource attribute that should otherwise be present

The only real downside is that this couples the docker container name
rather tightly to the service name for OTel. What would be really nifty
is if we could specify the value of `LS_SERVICE` (or the equivalent
`OTEL_` var as we switch over to native OTel SDKs rather than LS
launchers) based on the value of the docker container name! Or even
better, log natively with OTLP and include the service name there! Or
log with fluentd natively from services (or some other format) and
include the service name based on whatever the app says (LS_SERVICE,
etc).

Thats' a future todo though.